fix(sonar): resolve remaining 8 MINOR findings (const-correctness + cleanup)
The full CI Sonar rescan of the prior commit left 8 MINOR findings (gate
already green - 0 blocker/critical/major):

- S995 pointer-to-const: parse_packet_header + build_crypto_frame (qc),
  dtls_hrr_cookie_ok (c), gctr + gcm_core (ctx). Each only reads the pointee;
  the two GCM ones cascaded from the const aes256_ecb change. Const-qualified.
- S1905: drop the redundant (uint32_t) cast in failsafe json (uint32 subtraction
  is already uint32).
- S3628 "use a raw string literal": suppressed - escaped-quote JSON literals are
  the uniform style across src/ (196 instances, zero raw strings), same explicit
  stance as the index-loop / no-auto suppressions.

All affected envs build + test clean in WSL (ssh_aesgcm, quic_conn, dtls_conn,
failsafe, h3_server, ssh).

+++ src/services/failsafe/failsafe.cpp
@@ -151,7 +151,7 @@ int detws_failsafe_json_at(uint32_t now, char *out, size_t cap)
         fs_put(out, cap, &n, "\",\"overdue\":");
         fs_put(out, cap, &n, detws_lifeline_overdue(now, l.last_feed_ms, l.deadline_ms) ? "true" : "false");
         fs_put(out, cap, &n, ",\"age_ms\":");
-        fs_put_u32(out, cap, &n, (uint32_t)(now - l.last_feed_ms));
+        fs_put_u32(out, cap, &n, now - l.last_feed_ms);
         fs_put(out, cap, &n, ",\"deadline_ms\":");
         fs_put_u32(out, cap, &n, l.deadline_ms);
         fs_put(out, cap, &n, "}");
